Remove standalone FROST signing process logic.
This commit deletes `signing_process.rs` and re-integrates its responsibility into the `coordinator.rs` module while cleaning up redundant logic. Enhances code organization and brings the signing flow closer to the coordinator abstraction.
