+++ src/common/constants.rs
@@ -1,2 +1,5 @@
 /// Timeout for DKG operations in seconds
 pub const DKG_TIMEOUT_SECONDS: u64 = 120;
+
+/// Timeout for signing operations in seconds
+const SIGNING_TIMEOUT_SECONDS: u64 = 30;
